Changed the commands to find the repository info when executed
Instead of finding the repository info when the workspace is opened, the info is found at the point that the command is executed, allowing multiple repositories to exist inside the workspace. When a workspace is opened, it is scanned to see if it's in a repository, or if it contains repositories, and if it does, the command's context is enabled.

Fixes #20